+++ src/main/java/TestsObjects/BuyItemByExistUserSavedDetailedAndCC.java
@@ -106,7 +106,7 @@ public class BuyItemByExistUserSavedDetailedAndCC extends setup
     {
         Feed feed = new Feed(driver);
         Assert.assertTrue(feed.PlaceOrderButtonclicibilety(),"checkout button not clickable");
-        float total = (float) (pricecheckout*1.1);
+        int total = (int) (pricecheckout*1.1);
         this.totalprice = Float.parseFloat(feed.GetTotalPriceCheckOut().substring(1));
         Assert.assertEquals(this.totalprice,total);
         feed.ClickPlaceOrder();
